Ported to nopCommerce 4.00, now compiles against aspnet core
