+++ Nop.Plugin.Payments.eWay/Controllers/PaymenteWayController.cs
@@ -1,16 +1,29 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.IO;
 using System.Linq;
-using System.Web.Mvc;
+using System.Text;
+using Microsoft.AspNetCore.Mvc;
+using Nop.Core;
+using Nop.Core.Domain.Orders;
+using Nop.Core.Domain.Payments;
 using Nop.Plugin.Payments.eWay.Models;
-using Nop.Plugin.Payments.eWay.Validators;
+using Nop.Services;
 using Nop.Services.Configuration;
 using Nop.Services.Localization;
+using Nop.Services.Logging;
+using Nop.Services.Orders;
 using Nop.Services.Payments;
+using Nop.Services.Security;
+using Nop.Services.Stores;
 using Nop.Web.Framework.Controllers;
+using Nop.Web.Framework.Mvc.Filters;
+using Nop.Web.Framework.Security;
 
 namespace Nop.Plugin.Payments.eWay.Controllers
 {
+
     public class PaymenteWayController : BasePaymentController
     {
         private readonly ISettingService _settingService;
@@ -24,10 +37,10 @@ namespace Nop.Plugin.Payments.eWay.Controllers
             this._localizationService = localizationService;
             this._eWayPaymentSettings = eWayPaymentSettings;
         }
-        
-        [AdminAuthorize]
-        [ChildActionOnly]
-        public ActionResult Configure()
+
+        [AuthorizeAdmin]
+        [Area("Admin")]
+        public IActionResult Configure()
         {
             var model = new ConfigurationModel
             {
@@ -39,10 +52,12 @@ namespace Nop.Plugin.Payments.eWay.Controllers
             return View("~/Plugins/Payments.eWay/Views/Configure.cshtml", model);
         }
 
-        [HttpPost]
-        [AdminAuthorize]
-        [ChildActionOnly]
-        public ActionResult Configure(ConfigurationModel model)
+        [HttpPost, ActionName("Configure")]
+        [FormValueRequired("save")]
+        [AuthorizeAdmin]
+        [AdminAntiForgery]
+        [Area("Admin")]
+        public IActionResult Configure(ConfigurationModel model)
         {
             if (!ModelState.IsValid)
                 return Configure();
@@ -56,91 +71,9 @@ namespace Nop.Plugin.Payments.eWay.Controllers
             return Configure();
         }
 
-        [ChildActionOnly]
-        public ActionResult PaymentInfo()
-        {
-            var model = new PaymentInfoModel();
-            
-            //CC types
-            model.CreditCardTypes = new List<SelectListItem>
-            {
-                new SelectListItem { Text = "VISA", Value = "VISA" },
-                new SelectListItem { Text = "MASTERCARD", Value = "MASTER CARD" },
-                new SelectListItem { Text = "BANKCARD", Value = "BANK CARD" },
-                new SelectListItem { Text = "AMEX", Value = "AMEX" },
-                new SelectListItem { Text = "DINERS", Value = "DINERS" },
-                new SelectListItem { Text = "JCB", Value = "JCB" }
-            };
 
-            //years
-            for (var i = 0; i < 15; i++)
-            {
-                var year = Convert.ToString(DateTime.Now.Year + i);
-                model.ExpireYears.Add(new SelectListItem { Text = year, Value = year });
-            }
 
-            //months
-            for (var i = 1; i <= 12; i++)
-            {
-                var text = string.Format("{0:00}", i);
-                model.ExpireMonths.Add(new SelectListItem()
-                {
-                    Text = text,
-                    Value = i.ToString(),
-                });
-            }
 
-            //set postback values
-            var form = Request.Form;
-            model.CardholderName = form["CardholderName"];
-            model.CardNumber = form["CardNumber"];
-            model.CardCode = form["CardCode"];
-            var selectedCcType = model.CreditCardTypes.FirstOrDefault(x => x.Value.Equals(form["CreditCardType"], StringComparison.InvariantCultureIgnoreCase));
-            if (selectedCcType != null)
-                selectedCcType.Selected = true;
-            var selectedMonth = model.ExpireMonths.FirstOrDefault(x => x.Value.Equals(form["ExpireMonth"], StringComparison.InvariantCultureIgnoreCase));
-            if (selectedMonth != null)
-                selectedMonth.Selected = true;
-            var selectedYear = model.ExpireYears.FirstOrDefault(x => x.Value.Equals(form["ExpireYear"], StringComparison.InvariantCultureIgnoreCase));
-            if (selectedYear != null)
-                selectedYear.Selected = true;
 
-            return View("~/Plugins/Payments.eWay/Views/PaymentInfo.cshtml", model);
-        }
-
-        [NonAction]
-        public override IList<string> ValidatePaymentForm(FormCollection form)
-        {
-            var warnings = new List<string>();
-
-            //validate
-            var validator = new PaymentInfoValidator(_localizationService);
-            var model = new PaymentInfoModel()
-            {
-                CardholderName = form["CardholderName"],
-                CardNumber = form["CardNumber"],
-                CardCode = form["CardCode"],
-            };
-            var validationResult = validator.Validate(model);
-            if (validationResult.IsValid) return warnings;
-
-            warnings.AddRange(validationResult.Errors.Select(error => error.ErrorMessage));
-            return warnings;
-        }
-
-        [NonAction]
-        public override ProcessPaymentRequest GetPaymentInfo(FormCollection form)
-        {
-            var paymentInfo = new ProcessPaymentRequest
-            {
-                CreditCardType = form["CreditCardType"],
-                CreditCardName = form["CardholderName"],
-                CreditCardNumber = form["CardNumber"],
-                CreditCardExpireMonth = int.Parse(form["ExpireMonth"]),
-                CreditCardExpireYear = int.Parse(form["ExpireYear"]),
-                CreditCardCvv2 = form["CardCode"]
-            };
-            return paymentInfo;
-        }
     }
 }
