+++ Nop.Plugin.Payments.eWay/Description.txt
@@ -1,8 +1,8 @@
 ﻿Group: Payment methods
 FriendlyName: Credit Card
 SystemName: Payments.eWay
-Version: 2.10
-SupportedVersions: 3.90
+Version: 2.20
+SupportedVersions: 4.00
 Author: nopCommerce team
 DisplayOrder: 1
 FileName: Nop.Plugin.Payments.eWay.dll
