+++ Nop.Plugin.Payments.eWay/Models/ConfigurationModel.cs
@@ -1,17 +1,23 @@
-﻿using Nop.Web.Framework;
-using Nop.Web.Framework.Mvc;
+﻿using Microsoft.AspNetCore.Mvc.Rendering;
+using Nop.Web.Framework.Mvc.ModelBinding;
+using Nop.Web.Framework.Mvc.Models;
 
 namespace Nop.Plugin.Payments.eWay.Models
 {
     public class ConfigurationModel : BaseNopModel
     {
+        public int ActiveStoreScopeConfiguration { get; set; }
+
         [NopResourceDisplayName("Plugins.Payments.eWay.UseSandbox")]
         public bool UseSandbox { get; set; }
+        public bool UseSandbox_OverrideForStore { get; set; }
 
         [NopResourceDisplayName("Plugins.Payments.eWay.CustomerId")]
         public string CustomerId { get; set; }
+        public string CustomerId_OverrideForStore { get; set; }
 
         [NopResourceDisplayName("Plugins.Payments.eWay.AdditionalFee")]
         public decimal AdditionalFee { get; set; }
+        public decimal AdditionalFee_OverrideForStore { get; set; }
     }
 }
