+++ Nop.Plugin.Payments.eWay/Models/PaymentInfoModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
-using System.Web.Mvc;
-using Nop.Web.Framework;
-using Nop.Web.Framework.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Nop.Web.Framework.Mvc.ModelBinding;
+using Nop.Web.Framework.Mvc.Models;
 
 namespace Nop.Plugin.Payments.eWay.Models
 {
@@ -15,30 +15,24 @@ namespace Nop.Plugin.Payments.eWay.Models
         }
 
         [NopResourceDisplayName("Payment.SelectCreditCard")]
-        [AllowHtml]
         public string CreditCardType { get; set; }
         [NopResourceDisplayName("Payment.SelectCreditCard")]
         public IList<SelectListItem> CreditCardTypes { get; set; }
 
         [NopResourceDisplayName("Payment.CardholderName")]
-        [AllowHtml]
         public string CardholderName { get; set; }
 
         [NopResourceDisplayName("Payment.CardNumber")]
-        [AllowHtml]
         public string CardNumber { get; set; }
 
         [NopResourceDisplayName("Payment.ExpirationDate")]
-        [AllowHtml]
         public string ExpireMonth { get; set; }
         [NopResourceDisplayName("Payment.ExpirationDate")]
-        [AllowHtml]
         public string ExpireYear { get; set; }
         public IList<SelectListItem> ExpireMonths { get; set; }
         public IList<SelectListItem> ExpireYears { get; set; }
 
         [NopResourceDisplayName("Payment.CardCode")]
-        [AllowHtml]
         public string CardCode { get; set; }
     }
 }
