+++ Nop.Plugin.Payments.eWay/Views/ConfigureOld.cshtml
@@ -1,11 +1,13 @@
-﻿@model Nop.Plugin.Payments.eWay.Models.ConfigurationModel
-@inject Nop.Core.IWebHelper webHelper
-@{
-    Layout = "_ConfigurePlugin";
+﻿@{
+    Layout = "";
 }
+@model ConfigurationModel
+@using Nop.Plugin.Payments.eWay.Models;
+@using Nop.Web.Framework;
+@using (Html.BeginForm())
+{
 
-@await Component.InvokeAsync("StoreScopeConfiguration")
-<form asp-controller="PaymenteWay" asp-action="Configure" method="post">
+    @Html.AntiForgeryToken()
     <div class="panel-group">
         <div class="panel panel-default">
             <div class="panel-body">
@@ -29,37 +31,34 @@
 
                 <div class="form-group">
                     <div class="col-md-3">
-                        <nop-override-store-checkbox asp-for="UseSandbox_OverrideForStore" asp-input="UseSandbox" asp-store-scope="@Model.ActiveStoreScopeConfiguration" />
-                        <nop-label asp-for="UseSandbox" />
+                        @Html.NopLabelFor(model => model.UseSandbox)
                     </div>
                     <div class="col-md-9">
-                        <nop-editor asp-for="UseSandbox" />
-                        <span asp-validation-for="UseSandbox"></span>
+                        @Html.NopEditorFor(model => model.UseSandbox)
+                        @Html.ValidationMessageFor(model => model.UseSandbox)
                     </div>
                 </div>
 
                 <div class="form-group">
                     <div class="col-md-3">
-                        <nop-override-store-checkbox asp-for="CustomerId_OverrideForStore" asp-input="CustomerId" asp-store-scope="@Model.ActiveStoreScopeConfiguration" />
-                        <nop-label asp-for="CustomerId" />
+                        @Html.NopLabelFor(model => model.CustomerId)
                     </div>
                     <div class="col-md-9">
-                        <nop-editor asp-for="CustomerId" />
-                        <span asp-validation-for="CustomerId"></span>
+                        @Html.NopEditorFor(model => model.CustomerId)
+                        @Html.ValidationMessageFor(model => model.CustomerId)
                     </div>
                 </div>
 
                 <div class="form-group">
                     <div class="col-md-3">
-                        <nop-override-store-checkbox asp-for="AdditionalFee_OverrideForStore" asp-input="AdditionalFee" asp-store-scope="@Model.ActiveStoreScopeConfiguration" />
-                        <nop-label asp-for="AdditionalFee" />
+                        @Html.NopLabelFor(model => model.AdditionalFee)
                     </div>
                     <div class="col-md-9">
-                        <nop-editor asp-for="AdditionalFee" />
-                        <span asp-validation-for="AdditionalFee"></span>
+                        @Html.NopEditorFor(model => model.AdditionalFee)
+                        @Html.ValidationMessageFor(model => model.AdditionalFee)
                     </div>
                 </div>
-
+                
                 <div class="form-group">
                     <div class="col-md-3">
                         &nbsp;
@@ -72,4 +71,4 @@
             </div>
         </div>
     </div>
-</form>
+}
