--- Nop.Plugin.Payments.eWay/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FluentValidation" version="6.2.1.0" targetFramework="net451" />
-  <package id="Microsoft.AspNet.Mvc" version="5.2.3" targetFramework="net451" />
-  <package id="Microsoft.AspNet.Razor" version="3.2.3" targetFramework="net451" />
-  <package id="Microsoft.AspNet.WebPages" version="3.2.3" targetFramework="net451" />
-  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net451" />
-</packages>
